Refactor parseDate, isValidTaskTime. TaskTime -> DateParser
